+++ src/VtuneLaunchScriptWriter.ts
@@ -145,7 +145,7 @@ export class VtuneLaunchScriptWriter implements LaunchScriptWriter {
 	    await fs.promises.mkdir(parentFolder, { recursive: true });
 	    await fs.promises.writeFile(launchScriptPath, command, { mode: 0o744 });
 	    // vscode.window.showInformationMessage(command);
-	    vscode.commands.executeCommand('workbench.action.tasks.runTask', 'Intel(R) oneAPI: Launch VTune Profiler');
+	    vscode.commands.executeCommand('workbench.action.tasks.runTask', 'Intel® oneAPI: Launch VTune Profiler');
 	  }
   }
 }